improve db struct

+++ database/migrations/2022_04_07_074233_create_categories_table.php
@@ -15,7 +15,7 @@ class CreateCategoriesTable extends Migration
     {
         Schema::create('categories', function (Blueprint $table) {
             $table->id();
-            $table->integer('is_active');
+            $table->tinyInteger('is_active');
             $table->integer('parent_id');
             $table->string('title');
             $table->string('url')->unique();
