+++ database/migrations/2022_04_07_222321_create_blog_categories_table.php
@@ -15,7 +15,7 @@ class CreateBlogCategoriesTable extends Migration
     {
         Schema::create('blog_categories', function (Blueprint $table) {
             $table->id();
-            $table->integer('is_active');
+            $table->tinyInteger('is_active');
             $table->string('title');
             $table->string('url')->unique();
             $table->text('desc')->nullable();
